requirements.txt
aiogram==3.21.0
python-dotenv==1.1.1
requests==2.31.0
